requirements.txt
joblib==1.3.2
matplotlib==3.8.1
numpy==1.26.4
pandas==2.0.3
PyYAML==6.0.1
scanpy==1.10.3
scikit_learn==1.3.2
scipy==1.14.1
setuptools==65.5.0
torch==2.3.0
torch_geometric==2.6.1
